# openapi.yml
openapi: 3.1.0
info:
  title: Social Hours Platform API
  version: 0.1.0
  description: >
    Stack-neutral MVP API contract for a self-hosted social-hours platform.

    The contract includes core platform endpoints, Odoo-like module

    management endpoints, and dummy auth/student-data provider endpoints

    used to demonstrate integration with multiple college-owned systems.


    ## Student RBAC (whitelist)


    Users with only the `student` role may access a fixed allowlist of
    endpoints.

    All other `/api/v1/*` routes return 403. Staff roles (`admin`,
    `coordinator`,

    `faculty_supervisor`, `external_supervisor`, `auditor`) bypass the
    whitelist.


    | Method | Path | Student |

    |--------|------|---------|

    | GET | /api/v1/files/.+, /api/v1/me/?, /api/v1/me/applications,
    /api/v1/me/assignments, /api/v1/me/document-requirements,
    /api/v1/me/document-uploads, /api/v1/me/hour-logs, /api/v1/me/notifications,
    /api/v1/me/profile, /api/v1/me/social-services/history, /api/v1/projects/?,
    /api/v1/projects/{id} | yes (published only) |

    | POST | /api/v1/auth/introspect, /api/v1/auth/login,
    /api/v1/me/document-requirements/{id}/upload, /api/v1/me/evidence,
    /api/v1/me/hour-logs, /api/v1/me/profile/refresh,
    /api/v1/projects/{id}/applications/? | yes |

    | PATCH | /api/v1/me/notifications/{id}/read | yes |
servers:
  - url: https://social-hours.example.edu
    description: Core platform instance
...
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/setup/test:
    post:
      tags: [ Setup ]
      summary: Test configured connectors during setup
      operationId: testSetupConnectors
      security: []
      responses:
        "200":
          description: Connector test results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SetupConnectorTestResult"
        "410":
          description: Setup already completed
...
  /api/v1/modules/{moduleKey}/test:
    post:
      tags: [ Modules ]
      summary: Test module connection and contract readiness
      operationId: testModule
      parameters:
        - $ref: "#/components/parameters/ModuleKey"
      responses:
        "200":
          description: Module test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModuleTestResult"
...
  /api/v1/provider-connections/{providerId}/test:
    post:
      tags: [ Provider Connections ]
      summary: Test a provider connection
      operationId: testProviderConnection
      parameters:
        - $ref: "#/components/parameters/ProviderId"
      responses:
        "200":
          description: Test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderTestResult"
...
  /api/v1/applications:
    get:
      tags: [ Applications ]
      summary: List applications
      operationId: listApplications
      description: Students only see their own applications unless they also hold
        admin or coordinator role.
      parameters:
        - in: query
          name: status
          schema:
            $ref: "#/components/schemas/ApplicationStatus"
        - in: query
          name: studentRef
          schema:
            type: string
      responses:
        "200":
          description: Applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectApplication"
...
  /api/v1/assignments:
    get:
      tags: [ Assignments ]
      summary: List assignments
      operationId: listAssignments
      description: Students only see their own assignments unless they also hold
        admin, coordinator, or faculty_supervisor role.
      parameters:
        - in: query
          name: studentRef
          schema:
            type: string
        - in: query
          name: projectId
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assignment"
...
  /api/v1/hour-logs:
    get:
      tags: [ Hour Logs ]
      summary: List hour logs
      operationId: listHourLogs
      parameters:
        - in: query
          name: assignmentId
          schema:
            type: string
            format: uuid
        - in: query
          name: status
          schema:
            $ref: "#/components/schemas/ApprovalStatus"
      responses:
        "200":
          description: Hour logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HourLog"
    post:
      tags: [ Hour Logs ]
      summary: Submit an hour log
      operationId: createHourLog
      description: Requires student role. The assignment must belong to the
        authenticated student.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HourLogInput"
      responses:
        "201":
          description: Hour log created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HourLog"
  /api/v1/hour-logs/{hourLogId}/approve:
    post:
      tags: [ Hour Logs ]
      summary: Approve an hour log
      operationId: approveHourLog
      parameters:
        - $ref: "#/components/parameters/HourLogId"
      responses:
        "200":
          description: Hour log approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HourLog"
  /api/v1/hour-logs/{hourLogId}/reject:
    post:
      tags: [ Hour Logs ]
      summary: Reject an hour log
      operationId: rejectHourLog
      parameters:
        - $ref: "#/components/parameters/HourLogId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ reason ]
              properties:
                reason:
                  type: string
      responses:
        "200":
          description: Hour log rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HourLog"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    webhookSignature:
      type: apiKey
      in: header
      name: X-Webhook-Signature
  parameters:
    ModuleKey:
      in: path
      name: moduleKey
      required: true
      schema:
        type: string
      examples:
        dummyAuth:
          value: dummy-auth-connector
        dummyStudentData:
          value: dummy-student-data-connector
...
    ProviderId:
      in: path
      name: providerId
      required: true
      schema:
        type: string
        format: uuid
...
    HourLogId:
      in: path
      name: hourLogId
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    SetupConnectorTestResult:
      type: object
      required: [ results, allOk ]
      properties:
        allOk:
          type: boolean
        results:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              moduleKey:
                type: string
              ok:
                type: boolean
              message:
                type: string
...
    ModuleTestResult:
      type: object
      required: [ moduleKey, status, checkedAt ]
      properties:
        moduleKey:
          type: string
        status:
          type: string
          enum: [ success, failure ]
        contract:
          type: string
          examples: [ auth.v1, student_data.v1 ]
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        checkedAt:
          type: string
          format: date-time
...
    ProviderTestResult:
      type: object
      required: [ status, checkedAt ]
      properties:
        status:
          type: string
          enum: [ success, failure ]
        message:
          type: string
        checkedAt:
          type: string
          format: date-time
...
    ProjectApplication:
      type: object
      required: [ id, projectId, studentRef, status, createdAt ]
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        studentRef:
          type: string
        status:
          $ref: "#/components/schemas/ApplicationStatus"
        motivation:
          type: string
        rejectionReason:
          type: string
        createdAt:
          type: string
          format: date-time
    ApplicationStatus:
      type: string
      enum: [ submitted, approved, rejected, cancelled, waitlisted ]
    Assignment:
      type: object
      required: [ id, projectId, studentRef, status ]
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        studentRef:
          type: string
        supervisorRef:
          type: string
        status:
          type: string
          enum: [ active, completed, cancelled, suspended ]
    HourLogInput:
      type: object
      required: [ assignmentId, date, durationHours, category, description ]
      properties:
        assignmentId:
          type: string
          format: uuid
        date:
          type: string
          format: date
        startTime:
          type: string
        endTime:
          type: string
        durationHours:
          type: number
        category:
          type: string
          enum: [ disciplinary, environmental, community, research, administrative, other ]
        description:
          type: string
        evidenceIds:
          type: array
          items:
            type: string
            format: uuid
    HourLog:
      allOf:
        - $ref: "#/components/schemas/HourLogInput"
        - type: object
          required: [ id, status, createdAt ]
          properties:
            id:
              type: string
              format: uuid
            status:
              $ref: "#/components/schemas/ApprovalStatus"
            rejectionReason:
              type: string
            createdAt:
              type: string
              format: date-time
            evidence:
              type: array
              items:
                $ref: "#/components/schemas/HourLogEvidence"
    HourLogEvidence:
      type: object
      required: [ id, fileName, storageRef ]
      properties:
        id:
          type: string
          format: uuid
        fileName:
          type: string
        storageRef:
          type: string
    ApprovalStatus:
      type: string
      enum: [ pending, approved, rejected ]
...